added backend role-based-access

+++ server/api/donor/[id].delete.ts
@@ -4,7 +4,14 @@ const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) => {  
     try {
-        const id = getRouterParam(event, 'id');
+        const id = await getRouterParam(event, 'id');
+        const body = await readBody(event);
+        if(body.permissionLevel < 1){
+            throw createError({
+                statusCode:401,
+                statusMessage:"User does not have permission to delete donors"
+            })
+        }
         const deletedDonor = await prisma.donor.delete({
             where: { id }
         });
