+++ server/api/grants/[id].delete.ts
@@ -4,7 +4,14 @@ const prisma = new PrismaClient()
 
 export default defineEventHandler(async (event) => {
     try {
-        const grantId = getRouterParam(event, 'id');
+        const grantId = await getRouterParam(event, 'id');
+        const body = await readBody(event);
+        if(body.permissionLevel < 1){
+            throw createError({
+                statusCode:401,
+                statusMessage:"User does not have permission to delete grants"
+            })
+        }
         const grant = await prisma.grant.delete({
             where: {id:grantId},
         });
