+++ server/api/user/[id].delete.ts
@@ -5,7 +5,14 @@ const prisma = new PrismaClient();
 
 export default defineEventHandler(async (event) =>{
     try{
-        const id = getRouterParam(event, 'id');
+        const id = await getRouterParam(event, 'id');
+        const body = await readBody(event);
+        if(body.permissionLevel < 2){
+            throw createError({
+                statusCode:401,
+                statusMessage:"User does not have permission to delete users"
+            })
+        }
         const user = await prisma.user.delete({
             where:{
                 id:id
